refactor: extract logic into hook
